Fixed canvas resize issue
Top canvas resizing was using the bottom canvas height. This made it
so that the top canvas would have a smaller viewing area than what it
actually should of had. The issue was die to setting the height and
width of the top and bottom canvas at the same time in recalcLayerScale.