Updated file specification
File specification required that board shape contain all paths used
to define the board shape. This data was also replicated within the
layers section of the document. Removed board shape definition so that
only the bounding box for he design is provided.

With this change less redundant data is now needed.