Fixed rending issues with Vias
Vias were not being rendered correctly due to a bug when creating a via
object. The layer number was not being saved which caused iPCB to not
correctly draw vias.

Also fixed drill holes not showing up for square and circle vias. Issue
was that diameter was spelled wrong.

Removed rendering files no longer needed.